+++ mne_bids/__init__.py
@@ -8,4 +8,5 @@ from mne_bids.path import (BIDSPath, get_datatypes, get_entity_vals,
 from mne_bids.read import get_head_mri_trans, read_raw_bids
 from mne_bids.utils import (get_anonymization_daysback)
 from mne_bids.write import (make_dataset_description, write_anat,
-                            write_raw_bids, mark_bad_channels)
+                            write_raw_bids, mark_bad_channels,
+                            write_fine_calibration, write_cross_talk)